Added completion flag to the SynchronizedRequestStream
While the SynchronizedRequestStream took care of race conditions in calling onNext methods, it did not take into account that a stream could be concurrently completes and have messages sent to it. This would (if completion occurred first) still result in exceptions. This change causes any messages sent after completing a stream to be dropped.