Delay reconnect on certain disconnect reasons
When a connection fails with an error that one cannot recover from automatically, the retry interval should be extended. This is for example required when a call fails with a PERMISSION_DENIED. In such case, an attempt after 500ms is unlikely to succeed. The interval is extended to 5s in such cases.