Simplified connection management
The connection management and state change detection on these connections has been simplified. The responsibility for keeping a connection is now completely managed by the AxonServerManagedChannel. The Communication Channels are no longer involved in verifying these connection.

Logging of the Query Handler registration moved to info level, similar to the Command Handler registration, for consistency.