Fix an issue with slow reconnection when server nodes shut down
When shutting down a server node, application nodes attempt to recover the lost connection before setting up a new connection. Attempts to configure a connection timeout failed as they seem to be ignored by gRPC/Netty.

This commit ensures the connect timeout is also taken into account when attempting to recover a previously working connection. It will also ensure "forcePlatformReconnect" is respected in such case, forcing an immediate reconnect when a connection isn't in the ready state anymore.