Fix off-by-one issue in token-at-timestamp
When creating token based on a given timestamp, the stream using that token would skip the very first event at that timestamp. This commit fixes this issue, by converting the sequence to a token that results in the event with given sequence to be returned as first event in the stream.