Catch open stream exceptions to clean-up on failures
Whenever the CommandChannelImpl and QueryChannelImpl try to open a
stream with the CommandProviderOutbound and QueryProviderOutbound
respectively, we should catch exceptions. If we do not, the channel
implementations get stuck in a faulty state, where they created their
StreamObservers, but weren't able to connect them. This situation can
occur, if an AxonServerConnector-Java user dispatches a command/query
while all connections are down. By catching the exception, and clearing
the StreamObservers, we resolve this issue.

#bug/stream-cleanup-on-opening-failure